test(update): keep schema inspection coverage within shard limits
